update ai packages
remove unneeded dependencies, extract embeedings into its own model
