+++ packages/AI/Core/src/generic/embed.types.ts
@@ -1,8 +1,7 @@
-import { ModelUsage } from "./baseModel";
+import { BaseParams, ModelUsage } from "./baseModel";
 
-export type EmbedParams = {
-    text: string,
-    model: string
+export type EmbedParams = BaseParams & {
+    text: string
 }
 
 export type EmbedResult = {
