+++ packages/AI/Core/src/index.ts
@@ -6,3 +6,4 @@ export * from './generic/classify.types';
 export * from './generic/summarize.types';
 export * from './generic/apiKeyDictionary';
 export * from './generic/embed.types';
+export * from './generic/Embeddings';
