+++ packages/AI/Mistral/src/index.ts
@@ -1,1 +1,2 @@
 export * from './models/mistral';
+export * from './models/mistralEmbedding';
