+++ packages/AI/OpenAI/src/index.ts
@@ -1,1 +1,3 @@
 export * from './models/openAI';
+export * from './models/openAIEmbedding';
+export * from './models/embeddingModels.types';
