+++ packages/AI/VectorDupe/package.json
@@ -16,10 +16,8 @@
   "license": "ISC",
   "dependencies": {
     "@memberjunction/ai": "^0.9.159",
-    "@memberjunction/ai-mistral": "^0.9.16",
     "@memberjunction/ai-vectordb": "^0.9.3",
     "@memberjunction/ai-vectors": "^0.9.4",
-    "@memberjunction/ai-vectors-pinecone": "^0.9.7",
     "@memberjunction/aiengine": "^0.9.56",
     "@memberjunction/core": "^0.9.171",
     "@memberjunction/global": "^0.9.152",
