+++ packages/AI/VectorSync/src/generic/entity.types.ts
