+++ packages/AI/VectorSync/src/generic/entitySyncConfig.types.ts
