+++ packages/AI/VectorSync/src/index.ts
@@ -1,1 +1,4 @@
-import { EntityVectorSyncer } from './entityVectorSyncher';
+export * from './generic/entity.types';
+export * from './models/vectorSyncBase';
+export * from './generic/entitySyncConfig.types';
+export * from './models/entityVectorSync';
