+++ packages/AI/VectorSync/src/models/vectorSyncBase.ts
