--- packages/AI/VectorsEntitySync/.vscode/tasks.json
@@ -1,14 +1,0 @@
-{
-	"version": "2.0.0",
-	"tasks": [
-		{
-			"type": "typescript",
-			"tsconfig": "tsconfig.json",
-			"problemMatcher": [
-				"$tsc"
-			],
-			"group": "build",
-			"label": "tsc: build - tsconfig.json"
-		}
-	]
-}
