--- packages/AI/VectorsEntitySync/data/entitiesToSync.json
@@ -1,14 +1,0 @@
-[
-  {
-    "EntityDocumentID": 1,
-    "Interval": 3600,
-    "RunViewParams": {
-      "OrderBy": "ID",
-      "IgnoreMaxRows": true
-    },
-    "IncludeInSync": true,
-    "LastRunDate": "2024-03-04T18:50:40.225Z",
-    "VectorIndexID": 1,
-    "VectorID": 1
-  }
-]
