--- packages/AI/VectorsEntitySync/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "@memberjunction/ai-vectors-entity-sync",
-  "version": "0.9.6",
-  "description": "MemberJunction: AI Vector/Entity Sync Package - handles synchronization between Vector DB and MJ CDP Data",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "/dist"
-  ],
-  "scripts": {
-    "start": "ts-node-dev src/index.ts",
-    "build": "tsc",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "MemberJunction.com",
-  "license": "ISC",
-  "dependencies": {
-    "@memberjunction/ai": "^0.9.159",
-    "@memberjunction/ai-mistral": "^0.9.16",
-    "@memberjunction/ai-vectordb": "^0.9.3",
-    "@memberjunction/ai-vectors": "^0.9.4",
-    "@memberjunction/ai-vectors-pinecone": "^0.9.7",
-    "@memberjunction/aiengine": "^0.9.56",
-    "@memberjunction/core": "^0.9.171",
-    "@memberjunction/global": "^0.9.152",
-    "@memberjunction/sqlserver-dataprovider": "^0.9.191",
-    "dotenv": "^16.0.3",
-    "typeorm": "^0.3.20"
-  },
-  "devDependencies": {
-    "@types/mssql": "^9.1.5",
-    "@types/node": "^20.1.7",
-    "ts-node-dev": "^2.0.0",
-    "typescript": "^5.3.3"
-  }
-}
