--- packages/AI/VectorsEntitySync/src/index.ts
@@ -1,5 +1,0 @@
-export * from './generic/duplicateRecords.types';
-export * from './generic/entity.types';
-export * from './models/vectorSyncBase';
-export * from './generic/entitySyncConfig.types';
-export * from './models/entityVectorSyncer';
