--- packages/AI/VectorsEntitySync/tsconfig.json
@@ -1,25 +1,0 @@
-{
-    "compilerOptions": {
-      "module": "commonjs",
-      "target": "es2022",
-      "declaration": true,
-      "outDir": "dist",
-      "rootDir": "src",
-      "sourceMap": true,
-      "experimentalDecorators": true,
-      "resolveJsonModule": true,
-      "esModuleInterop": true,
-      "lib": ["es2022", "esnext.asynciterable"],
-      "moduleResolution": "node",
-      "removeComments": true,
-      "noImplicitThis": true,
-      "noImplicitReturns": true,
-      "noFallthroughCasesInSwitch": true,
-      "allowSyntheticDefaultImports": true,
-      "emitDecoratorMetadata": true,
-      "skipLibCheck": true
-    },
-    "exclude": ["node_modules"],
-    "include": ["./src/**/*.ts"]
-  }
-
