+++ packages/AI/VectorsPinecone/package.json
@@ -15,7 +15,6 @@
   "author": "MemberJunction.com",
   "license": "ISC",
   "dependencies": {
-    "@langchain/mistralai": "^0.0.7",
     "@memberjunction/aiengine": "^0.9.56",
     "@memberjunction/core": "^0.9.171",
     "@memberjunction/global": "^0.9.152",
@@ -27,7 +26,6 @@
     "typeorm": "^0.3.20"
   },
   "devDependencies": {
-    "@types/mssql": "^9.1.5",
     "@types/node": "^20.1.7",
     "ts-node-dev": "^2.0.0",
     "typescript": "^5.3.3"
